+++ any-adapter/src/main/java/de/pfaffenrodt/adapter/SimpleDiffCallback.kt
@@ -1,0 +1,11 @@
+package de.pfaffenrodt.adapter
+
+class SimpleDiffCallback : DiffCallback<Any>() {
+    override fun areItemsTheSame(oldItem: Any, newItem: Any): Boolean {
+        return oldItem == newItem
+    }
+
+    override fun areContentsTheSame(oldItem: Any, newItem: Any): Boolean {
+        return oldItem == newItem
+    }
+}