+++ buildSrc/src/main/java/Dependencies.kt
@@ -4,9 +4,17 @@ object Dependencies {
 
     object Android {
         const val appCompat = "androidx.appcompat:appcompat:1.2.0"
-        const val recyclerview = "androidx.recyclerview:recyclerview:1.1.0"
+        const val recyclerview = "androidx.recyclerview:recyclerview:${Version.recyclerviewVersion}"
         const val lifecycleRuntime = "androidx.lifecycle:lifecycle-runtime:${Version.lifecycleVersion}"
+
+        object Paging {
+            const val runtime = "androidx.paging:paging-runtime:${Version.pagingVersion}"
+            const val common = "androidx.paging:paging-common:${Version.pagingVersion}"
+            const val rxjava3 = "androidx.paging:paging-rxjava3:${Version.pagingVersion}"
+        }
     }
+
+    const val rxjava3 = "io.reactivex.rxjava3:rxjava:3.0.8"
 
     object Test {
         const val espresso = "androidx.test.espresso:espresso-core:3.3.0"