+++ sample-kotlin/src/main/java/de/pfaffenrodt/adapter/sample/presenter/SamplePresenterB.kt
@@ -1,0 +1,8 @@
+package de.pfaffenrodt.adapter.sample.presenter
+
+import de.pfaffenrodt.adapter.sample.R
+
+class SamplePresenterB : SamplePresenterA() {
+    override val layoutId: Int
+        get() = R.layout.item_sample_b
+}